Use placholder component to render baddie list
